Move Fake string table to test module
